# bindings/javascript: Reduce VM/native crossing overhead